Initiate the project
This patch implements the basic framework of binding the RCL library
to the JavaScript interface, meanwhile brings out the timer object and
with an example of it.

The project is using eslint (JavaScript) and cpplint (c/c++) to check
the static syntax. Hopefully, unit tests will be added later. And the
buildbot is running on travis-ci with docker srevice.