[rosidl_gen]Use ref to generate and initialize the string message
Before this patch, we bound the JS method to the native method to
initialize and assign the message of a string type. So the overhead
increased.

This patch is using the unicode of '\0' to initialize an empty string
and set the corresponding values of size/capacity, which will also
simplify the case when the message is a string array.

Fix #47