[windows][ci]Enable ci on windows platform
This patch enables the ci on windows platform. As the appveyor allows
a maximum of 60min build, which is not enough for our task, we choose to
download the latest ros binary package from http://ci.ros2.org, which is
often built several hours ago.

We build our rclnodejs addon based on the package from build farm. At
this moment, we will not trigger the test.

Fix