Support to pass primitive type directly to publisher/client
Currently, when we want to pass a primitive type of message to publisher
or client, we have to create a new class object of that type and assign
the value to its 'data' property, which makes the whole process redundant.

This patch lets user publish primitive message directly, the transition
from primitive message to the corresponding object will occur
implicitly. So you can write code like below:

  const publisher = node.createPublisher(String, 'topic');
  publisher.publish('hello world');

Besides, we also update the example to demonstrate the new usage for
this change.

Fix #50