Support to require a single type of message/service
Currently if we want to require a type of message, you have to get the
whole messages of that package, which will cost a lot of time because of
the sync manipulation.

e.g. If we want to require the String class, and write:
    // This gets all stuff included in the std_msgs package.
    const String = rclnodejs.require('std_msgs').msg.String;

To improve the performance, this patch enables to require a single type
of message which belongs to a specific package.

e.g. You can get the String class as follows:
    const String = rclnodejs.require('std_msgs/msg/String');

Fix #179