Temporarily diasble the cases for interaction test
As mentioned in issue 222
(https://github.com/RobotWebTools/rclnodejs/issues/222), we disable
these test cases on all ci platform. We will revert this patch, when the
issue is fixed.

This patch also added the dependencies for Rviz on circleci,
as we build ros2 from scratch on it.

Also, remove the node spacename for the cpp unit test.

What's more, a new dependency named "Rviz" has been added to the ROS2.0
build which will triger the travis to exceed the maximum build time, so
we decided to download the binary package to save the time.