Use mumber variable of message class to hold the array element
When a message which has an array property, user can get it by the get
method, meanwhile any element in the array can be re-assigned by
operator = through index. But the array gotten by the get method is generated as a
temporary variable and will not influence the one which is stored in
the original message.

This patch creates a member variable to hold the property, which has a
type of primitive array.

Fix #232