Check the consistency of the message's members when being serialized
Currently when a message is instantiated, the values of its data members will
be a "default". Reference the issue 243:

If we have a request which will be sent to the service to calculate the
sum of two integers, and create {a: 1} instead of {a: 1, b: 1} which
means we do not assign the value for member b explicitly. What the
service will receive is {a: 1, b: 0} from the client request. We may
just forget the assignment unintentionally, but rclnodejs should produce
an error.

This patch implements the strategy below:
1. If the type of a data member in the message is primitive, including
Bool, Int8, UInt8, Int16, UInt16, Int32, UInt32, Int64, UInt64, Float64,
Float32, Char, Byte and String, the data member will be not be created and
undefined is returned when being queried.

2. Other types of the data members, including compound data member and
arrays (primitive & non primitive), it will be created.

If the message is being serialized, which occurs during publishing a topic or
sending a request, the consistency of the message will be checked. Once
one of the data members is undefined, a TypeError will be thrown to
indicate the name of the invalid one.

Fix #243