[PnP]Add PnP tests for rclnodejs client (Python)
This patch added the tests used for rclnodejs. We defined two kinds of
PnP tests, which are stress testing and endurance testing.

1. Stress testing
The test case is designed to find the behaviour beyond the normal
expected overload, and we can find the breaking points and short term
memory leaks through the tests. There is one test case for each of the
publisher/subscription and client/service. The strategy is:

- A publisher publishes a topic of std_msgs/UInt8MultiArray type every
  100ms, which contains an array of size 10MB, and sends 36000 times
  totally.
- A client sends a request of nav_msgs/GetMap whose response will
  contains an array of size 10MB until it receives the response from the
  service 36000 times.

2. Endurance testing
The test case is designed to find the expected overload for a longer
term, and we can find the long term memory leaks through the tests.
There is one test for each of the publisher/subscription and
client/service. The stragegy is:

- A publisher publishes a topic of std_msgs/JointState type every
  100ms for 24 hours (864000 totally).
- A client sends a request of std_srvs/SetBool until it receives the
  response from the service 864000 times.

All tests begin with a line of output 'Begin' and end with 'End'.

Fix #278