Lock the handles throughout a loop in child thread
Currently we only locked the mutex when adding the handles to the
wait_set, and there is a race condition that the handles may be
destroyed during executing the rcl_wait() function.

Instead we must lock the mutex throughout a loop to ensure that all
handles are valid which means none of them is destroyed. The mutex can
be unlocked when we have cleared the handles from the wat_set.

Fix #301