[rosidl]Move the translator to rosidl-generator module
Currently, the translator, which can adapt a plain
JavaScript object to an object of ROS message, resides in the lib/
folder. It seems reasonable to move it to rosidl-generator and be
wrapped by the message class instead of exposing its interfaces to
rclnodejs.

This patch implemented the function described above.

Fix NONE