Use dlerror() to get the error diagnostic
Currently if the library(.so/.dll) can not be found by the dlopen function,
a nullptr will be returned which will be passed to the functions of rcl
later, which is not reasonable.

This patch implements using the dlerror() to get the error message and
throw an error to JavaScript side which is catchable.

Fix #338