[rosidl]Support to assign the property of the message with a plain object
For current messages generated by rclnodejs, if you want to assign a
property of the message to a new value, the type of it must be the same
with the type of the property to be assigned, e.g.

let header = new Header();
let stamp = new Time({sec: 100, nanosec:100});
header.stamp = stamp;

This patch implemented that a plain object can be assign to the property
of the message, e.g.

let header = new Header();
header.stamp = {sec: 100, nanosec:100};

Fix #351