Correct the memory layout of fixed array
If the .msg contains a field of array which has a fixed size, the struct
of it is different from the one which has a variable array. The struct
holds the pointer of the array when it's variable, and holds the array
itself directly when it's fixed. e.g.

/// Struct of message geometry_msgs/PoseWithCovariance
typedef struct geometry_msgs__msg__PoseWithCovariance
{
  geometry_msgs__msg__Pose pose;
    double covariance[36];

} geometry_msgs__msg__PoseWithCovariance;

The change of this patch differentiates the arrays which have fixed
size to have the right memory layout for primitive type, and patch for
non-primitive arrays of fixed size will be following.

Fix #400