commit log
* generate type class strings for messages and services
* narrow down `TypeClass` type from `string` to known values
* generate typings for *Wrapper javascript classes
* generate constants as 'const enum'

These provides intellisense for typeclass names and allow constants to be referenced by value.

The constants are generated with '_Constants' suffix, since ros does not allow underscores in
the message name, this should be safe from any conflicts. This is also same as how the ros idl
generator exports the constants.