Add actions to the event loop (#570)
Adds action clients and servers into the event loop. You can't add action client/servers right now so this change essentially does nothing.

Actions present an interesting problem where not only do you need to know if the handle is ready, but you also need to know what to execute within the handle. rclpy abstracts actions with a Waitable base class and lets the sub-class fill in the details of reading from the wait_set. This approach doesn't fit in with the design of rclnodejs, so instead actions are treated as entities, similar to publishers, subscribers, etc, but with some special handling.

To know what part of the action server/client should be processed a property bag has added to RclHandle allowing the event loop to attach arbitrary data to be sent back to the main thread. The property bag could be extended in the future if ever needed by other entity types.

Below is how the property bag would be used in later PRs.

```javascript
// From Node class
execute(handles) {
  let actionServersReady = this._actionServers.filter((actionServer) =>
      handles.indexOf(actionServer.handle) !== -1);

  // ...

  actionServersReady.forEach((actionServer) => {
    // Read the properties from the handle
    let properties = actionServer.handle.properties;

    if (properties.isGoalRequestReady) {
      // process goal request
    }

    if (properties.isCancelRequestReady) {
      // process cancel request
    }

    // process other requests
  }
}
```

Note this PR is targeting the 'actions' branch.