Remove spdlog as dependency
Currently, we depend on spdlog to log for C++ codes when debug build.
The reason is that there was no logging util in the early stage.

This patch uses the logging util from rcutils to replace spdlog.
Thus, we can remove the spdlog as a dependency and reduce
the size of npm package also.

Fix #634