fix: buffer overflows (#684)
Fix leak in CreateNode/ShadowNode

For some reason, assigning the object wrapping `RclHandle` to a new persistent handle causes the weak callback to not get called, and the wrapped `RclHandle` is lost. My theory is that now that it has a strong persistent handle keeping it alive, v8 decides that the weak persistent handle is not needed anymore and disposes it. The fix is to simply remove the `handle` accessors, since the `RclHandle`'s life is now handled by v8, it should be automatically cleaned up when there is no more reference to the `ShadowNode`.

* add cpp test artifacts to .gitignore

* fix possible buffer overflows and use after stack scope

* add helper scripts and instructions to run tests with asan

* fix cpplint errors

* make node.handle readonly; remove corrupted handle destruction test

Fix #674