Block the background thread when no handles are added
Currently, if parameter service is not enabled and no handles are added
into the wait set, the background thread will loop without doing
anything, which causes a high CPU load.

This patch implements that when there is no handles are attached to the
current node, the background thread will be blocked by a semaphore and
the main thread will signal it later when new handle has been created.

Fix #752