Disabled tests with union types due to DTSLint ExpectType issue
The $ExpectType parser inconsistently produces false negative reports on
the ordering of union types. This is most annoying and I don't believe
can be resolved by re-organizing union types. My understanding of the
issue is the union type ordering/inferencing rules are still a
work-in-progress. Thus, disabled tests with union types.