improve destroyed entity handling to avoid segfault
entity.js
- added static gcHandles()

node.js
- revised execute() to use for-of loops for each entity type. This
simplifies the looping of each group of entities through the use of
loop interruption (e.g., use continue statement) to detect a
destroyed entity before processing. Additionally it the for-of loop
reduces the number of isDestroyed() calls Lastly, added a call to
Entity.gcHandles() prior to exiting the method.