Catch the error when loading library failed (#1071)
Currently, `interfaces.d.ts`, a declaration file for ROS2 messages, will be generated when running `npm install`, which needs to require the `.js` files of messages and the `ref-napi` C++ addon gets loaded.

There could be a corner case that the `ref-napi` C++ addon is cross-compiled for arm arch, so loading `binding.node` dynamically failed on x86 platform, see details #1066.

This patch catches the error when loading the library failed, and if it happens, we will use a prebuild `interfaces.d.ts` instead. Meanwhiles, this patch moves out generating `interfaces.d.ts` file out of `generate-messages` command because it already executed.

Fix: #1066